requirements.txt
Flask==2.3.2
flask-cors==4.0.0
google-generativeai==0.8.0
vt-py==0.11.0
python-dotenv==1.0.0
